Handle dependencies via Pkg extensions
