+++ src/conversion.jl
@@ -19,38 +19,7 @@ Base.convert(::Type{Array}, A::LinearMap) = convert(Matrix, A)
 Base.convert(::Type{AbstractMatrix}, A::LinearMap) = AbstractMatrix(A)
 Base.convert(::Type{AbstractArray}, A::LinearMap) = convert(AbstractMatrix, A)
 
-# sparse: create sparse matrix representation of LinearMap
-function SparseArrays.sparse(A::LinearMap{T}) where {T}
-    M, N = size(A)
-    rowind = Int[]
-    nzval = T[]
-    colptr = Vector{Int}(undef, N+1)
-    v = fill(zero(T), N)
-    Av = Vector{T}(undef, M)
-
-    @inbounds for i in eachindex(v)
-        v[i] = one(T)
-        _unsafe_mul!(Av, A, v)
-        js = findall(!iszero, Av)
-        colptr[i] = length(nzval) + 1
-        if length(js) > 0
-            append!(rowind, js)
-            append!(nzval, Av[js])
-        end
-        v[i] = zero(T)
-    end
-    colptr[N+1] = length(nzval) + 1
-
-    return SparseMatrixCSC(M, N, colptr, rowind, nzval)
-end
-Base.convert(::Type{SparseMatrixCSC}, A::LinearMap) = sparse(A)
-SparseArrays.SparseMatrixCSC(A::LinearMap) = sparse(A)
-
 # special cases
-
-# ScaledMap
-SparseArrays.sparse(A::ScaledMap{<:Any, <:Any, <:VecOrMatMap}) =
-    A.λ * sparse(A.lmap.lmap)
 
 # UniformScalingMap
 Base.convert(::Type{AbstractMatrix}, J::UniformScalingMap) = Diagonal(fill(J.λ, J.M))
@@ -61,19 +30,10 @@ Base.convert(::Type{T}, A::WrappedMap) where {T<:Matrix} = convert(T, A.lmap)
 Base.Matrix{T}(A::VectorMap) where {T} = copyto!(Matrix{eltype(T)}(undef, size(A)), A.lmap)
 Base.convert(::Type{T}, A::VectorMap) where {T<:Matrix} = T(A)
 Base.convert(::Type{AbstractMatrix}, A::WrappedMap) = convert(AbstractMatrix, A.lmap)
-SparseArrays.sparse(A::WrappedMap) = sparse(A.lmap)
-Base.convert(::Type{SparseMatrixCSC}, A::WrappedMap) = convert(SparseMatrixCSC, A.lmap)
 
 # TransposeMap & AdjointMap
 for (T, t) in ((AdjointMap, adjoint), (TransposeMap, transpose))
     @eval Base.convert(::Type{AbstractMatrix}, A::$T) = $t(convert(AbstractMatrix, A.lmap))
-    @eval SparseArrays.sparse(A::$T) = $t(convert(SparseMatrixCSC, A.lmap))
-end
-
-# LinearCombination
-function SparseArrays.sparse(ΣA::LinearCombination{<:Any, <:Tuple{Vararg{VecOrMatMap}}})
-    mats = map(A->getfield(A, :lmap), ΣA.maps)
-    return sum(sparse, mats)
 end
 
 # CompositeMap
@@ -99,50 +59,19 @@ function Base.Matrix{T}(AB::CompositeMap{<:Any, <:Tuple{VecOrMatMap, VecOrMatMap
     B, A = AB.maps
     return mul!(Matrix{T}(undef, size(AB)), A.lmap, B.lmap)
 end
-function SparseArrays.sparse(AB::CompositeMap{<:Any, <:Tuple{VecOrMatMap, VecOrMatMap}})
-    B, A = AB.maps
-    return sparse(A.lmap)*sparse(B.lmap)
-end
 function Base.Matrix{T}(λA::CompositeMap{<:Any, <:Tuple{VecOrMatMap, UniformScalingMap}}) where {T}
     A, J = λA.maps
     return mul!(Matrix{T}(undef, size(λA)), J.λ, A.lmap)
 end
-function SparseArrays.sparse(λA::CompositeMap{<:Any, <:Tuple{VecOrMatMap, UniformScalingMap}})
-    A, J = λA.maps
-    return J.λ*sparse(A.lmap)
-end
 function Base.Matrix{T}(Aλ::CompositeMap{<:Any, <:Tuple{UniformScalingMap, VecOrMatMap}}) where {T}
     J, A = Aλ.maps
     return mul!(Matrix{T}(undef, size(Aλ)), A.lmap, J.λ)
-end
-function SparseArrays.sparse(Aλ::CompositeMap{<:Any, <:Tuple{UniformScalingMap, VecOrMatMap}})
-    J, A = Aλ.maps
-    return sparse(A.lmap)*J.λ
-end
-
-# BlockMap & BlockDiagonalMap
-function SparseArrays.sparse(A::BlockMap)
-    return hvcat(
-        A.rows,
-        convert(SparseMatrixCSC, first(A.maps)),
-        convert.(AbstractArray, _tail(A.maps))...
-    )
-end
-Base.convert(::Type{AbstractMatrix}, A::BlockDiagonalMap) = sparse(A)
-function SparseArrays.sparse(A::BlockDiagonalMap)
-    return blockdiag(convert.(SparseMatrixCSC, A.maps)...)
 end
 
 # KroneckerMap & KroneckerSumMap
 Base.Matrix{T}(A::KroneckerMap) where {T} = kron(convert.(Matrix{T}, A.maps)...)
 Base.convert(::Type{AbstractMatrix}, A::KroneckerMap) =
     kron(convert.(AbstractMatrix, A.maps)...)
-function SparseArrays.sparse(A::KroneckerMap)
-    return kron(
-        convert(SparseMatrixCSC, first(A.maps)),
-        convert.(AbstractMatrix, _tail(A.maps))...
-    )
-end
 
 function Base.Matrix{T}(L::KroneckerSumMap) where {T}
     A, B = L.maps
@@ -154,12 +83,6 @@ function Base.convert(::Type{AbstractMatrix}, L::KroneckerSumMap)
     A, B = L.maps
     IA = Diagonal(ones(Bool, size(A, 1)))
     IB = Diagonal(ones(Bool, size(B, 1)))
-    return kron(convert(AbstractMatrix, A), IB) + kron(IA, convert(AbstractMatrix, B))
-end
-function SparseArrays.sparse(L::KroneckerSumMap)
-    A, B = L.maps
-    IA = sparse(Diagonal(ones(Bool, size(A, 1))))
-    IB = sparse(Diagonal(ones(Bool, size(B, 1))))
     return kron(convert(AbstractMatrix, A), IB) + kron(IA, convert(AbstractMatrix, B))
 end
